Add :key option to session config example.
I have experienced strange behavior with the flash in Rails 4.1.1 when not defining the `:key` on the `config.session_store`. The strange behavior is the flash not being set, lost, or not appearing in the correct context (not the following action but one sometimes 2 actions later).

I've yet to pinpoint _why_ this is happening and will open an issue/pr when I do.