Use a test setup similar to actiondispatch's Cookie Store tests
I took another look at the changes we had to make in #220 and compared
that to how things work in `actiondispatch`'s Cookie Store tests and
realized we were close, but missing how we were setting the options for
our store. So I've revamped the setup to be in line with what's in
Rails' own tests. This should make it easier to keep things in line and
working, going forward.

We also need to patch in support for with_routing for integration tests
which was added in Rails 7.2. So we do that, only when necessary. We can
drop that patch when we drop Rails 7.1 support.

See the following two files for how Cookie Store tests (and their base
`ActionDispatch::IntegrationTest`) work today (as of Rails 8.0):
   * https://github.com/rails/rails/blob/8-0-stable/actionpack/test/dispatch/session/cookie_store_test.rb
   * https://github.com/rails/rails/blob/8-0-stable/actionpack/test/abstract_unit.rb