Default to request's cookies_same_site_protection option
This brings the ActiveRecordStore in line with the CookieStore that
ships with Rails. (see: https://github.com/rails/rails/pull/45501)

`ActionDispatch::Session::ActiveRecordStore` passes along whatever
options it was configure with, and by default that DOES NOT include a
`:same_site` value. So when `Rack::Session::SessionId` is created, it's
defaulting `:same_site` to `nil` because the option is missing. That means,
by the time `ActionDispatch`'s cookie middleware runs, there is a
`:same_site` key, so it won't set the default.